+++ src/main/java/com/wenxt/claims/repository/LtClaimChargesRepository.java
@@ -6,6 +6,6 @@ import org.springframework.stereotype.Repository;
 import com.wenxt.claims.model.LT_CLAIM_CHARGES;
 
 @Repository
-public interface LtClaimChargesRepository extends JpaRepository<LT_CLAIM_CHARGES, Long> {
+public interface LtClaimChargesRepository extends JpaRepository<LT_CLAIM_CHARGES, Integer> {
 
 }
